Phases 9-10: presets, RANDOM, full state; engine fixes found by the preset audit
Presets and state
- 46 factory presets in 12 categories (all 25 names from the spec), tuned
  with an exact inverse node-tuning helper (radiusForRatio)
- PresetManager: factory + user XML presets, favourites, prev/next,
  modified detection, host programs; performance/setup parameters
  (master, polyphony, bend, MPE, quality, freeze) are never changed
- Musical RANDOM: gentle mutation / SHIFT regeneration from weighted,
  material-aware distributions; reseeds CHAOS
- Complete DAW state: parameters + gestures + seed + preset metadata +
  version; restored instance renders bit-identically; legacy/hostile
  state handled; gestures reach the audio thread via per-node serials
  and a never-blocking try-lock
- Gestures: angles wrap and close modulo whole turns (drag around the
  core = seamless orbit), exact 9-digit serialisation, procedural
  orbit/sway gestures; synced drift phase-locked to the host timeline

Engine defects found by rendering every preset
- Coupling-compensation limit cycle near unison (audio-rate loop FM that
  pumped energy): static coincidence taper scaled by pair coupling,
  under-relaxation 0.25, authority +-12 %; regression test added
- BOW/AIR feedback forces were injected into outer nodes (non-collocated,
  active): a resting bow made METAL node B self-oscillate. Now CORE-only;
  only AIR turbulence spreads
- Bow stick region relative to bow velocity: every pressure x speed
  speaks on every material (speed = level, pressure = brightness)
- Wolf-tone limit: driven voices soft-limit edge rotations so AIR/BOW
  keep speaking at high COUPLING
- Sounding voices no longer follow the global EXCITER switch

Also: zero compiler warnings in the full plugin build, MOTION_SYSTEM.md,
playability maps, docs updates. 71 tests / 823 checks pass; pluginval
strictness 5 passes (GUI tests pending the editor).

+++ Source/Nonlinear/ChaosEngine.h
@@ -37,7 +37,7 @@ public:
     {
         const float c = dsp::clamp (chaos, 0.0f, 1.0f);
         const float a = c * c;
-        const double rate = 0.15 + 2.5 * c;
+        const double rate = 0.15 + 2.5 * static_cast<double> (c);
         for (size_t i = 0; i < detune.size(); ++i)
         {
             const float w = detune[i].next (dt, rate * (1.0 + 0.23 * static_cast<double> (i)));
